chore(pakage): janggi 최상위 패키지 생성
