+++ src/main/java/janggi/domain/board/setup/LeftSetUp.java
@@ -1,14 +1,14 @@
-package domain.board.setup;
+package janggi.domain.board.setup;
 
-import domain.board.Point;
-import domain.piece.Elephant;
-import domain.piece.Horse;
-import domain.piece.Piece;
-import domain.side.Side;
+import janggi.domain.board.Point;
+import janggi.domain.piece.Elephant;
+import janggi.domain.piece.Horse;
+import janggi.domain.piece.Piece;
+import janggi.domain.side.Side;
 import java.util.HashMap;
 import java.util.Map;
 
-public class OutSetUp implements BoardSetUp{
+public class LeftSetUp implements BoardSetUp{
     @Override
     public Map<Point, Piece> generate(Side side) {
         Map<Point, Piece> leftSetup = new HashMap<>();
@@ -16,15 +16,15 @@ public class OutSetUp implements BoardSetUp{
         if (side == Side.CHO){
             leftSetup.put(new Point(0, 1), new Elephant(side));
             leftSetup.put(new Point(0, 2), new Horse(side));
-            leftSetup.put(new Point(0, 6), new Horse(side));
-            leftSetup.put(new Point(0, 7), new Elephant(side));
+            leftSetup.put(new Point(0, 6), new Elephant(side));
+            leftSetup.put(new Point(0, 7), new Horse(side));
         }
 
         if(side == Side.HAN){
             leftSetup.put(new Point(9, 7), new Elephant(side));
             leftSetup.put(new Point(9, 6), new Horse(side));
-            leftSetup.put(new Point(9, 2), new Horse(side));
-            leftSetup.put(new Point(9, 1), new Elephant(side));
+            leftSetup.put(new Point(9, 2), new Elephant(side));
+            leftSetup.put(new Point(9, 1), new Horse(side));
         }
         return leftSetup;
     }
