+++ src/main/java/janggi/domain/piece/Chariot.java
@@ -1,14 +1,14 @@
-package domain.piece;
+package janggi.domain.piece;
 
-import domain.board.Board;
-import domain.board.Point;
-import domain.side.Side;
+import janggi.domain.board.Board;
+import janggi.domain.board.Point;
+import janggi.domain.side.Side;
 import java.util.List;
 
-public class Horse extends Piece {
-    private static final PieceName NAME = PieceName.HORSE;
+public class Chariot extends Piece {
+    private static final PieceName NAME = PieceName.CHARIOT;
 
-    public Horse(Side side) {
+    public Chariot(Side side) {
         super(NAME, side);
     }
