+++ src/main/java/janggi/domain/piece/Empty.java
@@ -1,7 +1,7 @@
-package domain.piece;
+package janggi.domain.piece;
 
-import domain.board.Board;
-import domain.board.Point;
+import janggi.domain.board.Board;
+import janggi.domain.board.Point;
 import java.util.List;
 
 public class Empty extends Piece{
