+++ src/main/java/janggi/domain/piece/Horse.java
@@ -1,14 +1,14 @@
-package domain.piece;
+package janggi.domain.piece;
 
-import domain.board.Board;
-import domain.board.Point;
-import domain.side.Side;
+import janggi.domain.board.Board;
+import janggi.domain.board.Point;
+import janggi.domain.side.Side;
 import java.util.List;
 
-public class Cannon extends Piece {
-    private static final PieceName NAME = PieceName.CANNON;
+public class Horse extends Piece {
+    private static final PieceName NAME = PieceName.HORSE;
 
-    public Cannon(Side side) {
+    public Horse(Side side) {
         super(NAME, side);
     }
